feat: add AVX-512BW/VL/VBMI/VBMI2/VNNI and F16C ISA support
Expand VOLK's ISA and machine hierarchy to cover the AVX-512 sub-ISAs
and F16C that were missing from archs.xml / machines.xml, and widen the
arch bitmask from int to uint64_t so the dispatcher can accommodate the
enlarged flag space without overflow.

### gen/archs.xml
- Add `f16c` arch (compiler flags: -mf16c / /arch:AVX)
- Add `avx512bw`, `avx512vl`, `avx512vnni`, `avx512vbmi`, `avx512vbmi2`
  arches with appropriate compiler flags and 64-byte alignment

### gen/machines.xml
- Add machine entries for `avx512bw`, `avx512vl`, `avx512vbmi`,
  `avx512vnni`, `avx512vbmi2` with correct cumulative arch chains
- Fix `avx512dq` machine to include `avx512bw` as a prerequisite

### lib/volk_rank_archs.h / .c
- Change `impl_deps` parameter type from `int*` to `uint64_t*` to
  support more than 32 simultaneous ISA flags
- Use `int64_t` for best-value accumulators to match
- Add early-return guard when `n_impls == 0`
- Improve fallback logic: search for "generic" by name before
  returning index 0; emit more informative warning messages
- Replace `strncmp(..., 20)` with `strcmp` for exact name matching

### tmpl/
- Propagate `uint64_t impl_deps` through all code-generation templates
  (volk.tmpl.c/h, volk_cpu.tmpl.c/h, volk_machine_xxx.tmpl.c,
  volk_machines.tmpl.h)

### lib/CMakeLists.txt
- Add `overrule_arch(avx512bw ...)` guard for non-x86 targets

+++ lib/volk_rank_archs.h
@@ -11,6 +11,7 @@
 #define INCLUDED_VOLK_RANK_ARCHS_H
 
 #include <stdbool.h>
+#include <stdint.h>
 #include <stdlib.h>
 
 #ifdef __cplusplus
@@ -22,12 +23,12 @@ int volk_get_index(const char* impl_names[], // list of implementations by name
                    const char* impl_name     // the implementation name to find
 );
 
-int volk_rank_archs(const char* kern_name,    // name of the kernel to rank
-                    const char* impl_names[], // list of implementations by name
-                    const int* impl_deps,     // requirement mask per implementation
-                    const bool* alignment,    // alignment status of each implementation
-                    size_t n_impls,           // number of implementations available
-                    const bool align          // if false, filter aligned implementations
+int volk_rank_archs(const char* kern_name,       // name of the kernel to rank
+                    const char* impl_names[],    // list of implementations by name
+                    const uint64_t* impl_deps,   // requirement mask per implementation
+                    const bool* alignment,       // alignment status of each implementation
+                    size_t n_impls,              // number of implementations available
+                    const bool align             // if false, filter aligned implementations
 );
 
 #ifdef __cplusplus
